+++ EasyPost/BetaFeatures/Parameters/CustomsInfo/Create.cs
@@ -40,6 +40,10 @@ namespace EasyPost.BetaFeatures.Parameters.CustomsInfo
         [NestedRequestParameter(typeof(Shipments.Create), Necessity.Optional, "customs_signer")]
         public string? CustomsSigner { get; set; }
 
+        [TopLevelRequestParameter(Necessity.Optional, "customs_info", "declaration")]
+        [NestedRequestParameter(typeof(Shipments.Create), Necessity.Optional, "declaration")]
+        public string? Declaration { get; set; }
+
         [TopLevelRequestParameter(Necessity.Optional, "customs_info", "eel_pfc")]
         [NestedRequestParameter(typeof(Shipments.Create), Necessity.Optional, "eel_pfc")]
         public string? EelPfc { get; set; }